python/koompi-ai/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "koompi-ai"
version = "0.1.0"
description = "AI integration for KOOMPI OS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "KOOMPI Team", email = "dev@koompi.org" }
]
dependencies = [
    "llama-cpp-python>=0.2.0",
    "openai-whisper>=20231117",
    "anthropic>=0.18.0",
    "numpy>=1.26.0",
    "sounddevice>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
